% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davePDFSlideStyle.R
\name{davePDFSlideStyle}
\alias{davePDFSlideStyle}
\title{Doc template for presentations}
\usage{
davePDFSlideStyle(template = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{template}{Which template to use}

\item{pandoc_args}{Any additional pandoc arguments (in a list)}

\item{...}{Arguments for `rmarkdown::beamer_presentation`}
}
\description{
Doc template for presentations
}

