% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daveRevealSlideStyle.R
\name{daveRevealSlideStyle}
\alias{daveRevealSlideStyle}
\title{Doc template for presentations}
\usage{
daveRevealSlideStyle(css = NULL, options = NULL, ...)
}
\arguments{
\item{css}{Stylesheet}

\item{options}{Options to pass Reveal}

\item{...}{Other parameters for}
}
\description{
Doc template for presentations
}

