% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daveSlideStyle.R
\name{daveSlideStyle}
\alias{daveSlideStyle}
\title{Doc template for presentations}
\usage{
daveSlideStyle(css = NULL, ...)
}
\arguments{
\item{css}{Stylesheet}

\item{...}{Other parameters for}
}
\description{
Doc template for presentations
}

